[OB-2718] refac: maintenance window simplification
In which we remove the discrete `IsInsideMaintenanceWindow` endpoint
from `MaintenanceSystem` and instead a new function to `MaintenanceInfo`
allowing applications to query whether they are inside the window.

This approach means that getting the set of windows and querying for the
latest one both follow the same code paths with respect to service
queries, which means less code to maintain and zero extra query
overhead.