[NT-0] doc: reuploaded readme gif
Making the repo public has broken all GitHub-hosted assets that we've uploaded. Fixing this first and most important one by reuploading the gif again.

I also took the liberty of removing the WIP section of the readme, as the page it links to is empty. We can revive it when content exists there.