[OF-1077] feat: Add ticket redemption
Adds the ability to redeem event tickets by submitting them to the
CHS endpoint.

Note that this cannot be automatically tested since it requires
events and tickets to be setup in the 3rd party vendor and requires
human interaction during the test. A test is provided with
instructions but is disabled by default.