[OB-2804] fix: Handle non-JSON error responses
This adds a function to HttpPayload to check if the payload has a
JSON mime-type. This is then used when printing error messages to
check if an error response is JSON formatted before trying to parse
it as JSON. If it is not JSON then it is simply printed as is.