[OF-1151] feat!: Add email link for user invite
This changes when there is an array of InviteUserRoleInfo to use a
new InviteUserRoleInfoCollection. This collection class now includes
an EmailLinkUrl parameter that can be provided to specify where to
redirect a user back to. It is OK to leave this blank. This applies
to the CreateSpace, CreateSpaceWithBuffer and BulkInviteToSpace
calls.

The same thing is also added as a new optional parameter to the
InviteToSpace call.