[OB-3040] fix: Resolving a QuotaSystem C# crash
A bug in the CSP wrapper generator means that reference-to-enum
parameters will be considered to be int pointers at the C# wrapper
level.

These int pointers are then passed to the generated C wrapper which
will dereference them before passing them to CSP's C++ API.

Those dereferences were the cause of the crash.

The fix here is to alter the signature of the QuotaSystem API such that
it passes enum params by value rather than reference.

A separate issue for the wrapper generator bug will be raised and
tracked.