[OB-3042] fix: quota queries for non-basic tiers
We had a typo in the strings we created to form URL GET requests,
where we were creating tier strings that began with an uppercase.

The convention here is for tier strings to be all lowercase.

The impact was that services could return 200 responses to the
incorrectly formatted GET requests but with the wrong information.

And the fix is to ensure tier string parsing is consistently treating
the tier names as all lower-case.