[OB-2323] fix: CSP crashing on M1 macs
Similar to recent crashes that have been fixed for
Quest & Android, mimalloc crashes when CSP is
initialized within an application running on an M1 Mac.

The short-term fix for this is the same as was applied for Quest.
We fall back to STL allocator behavior and avoid the usage of
mimalloc entirely.