[OB-3069] refac!: Expose multiplayer errors
This change addresses a lack of error handling in our multiplayer code.
A lot of calls return values were being ignored, which resulted in
internal errors that weren't properly handled. It also exposes a way
for the user to handle errors.

BREAKING CHANGE: This changes a lot of `MultiplayerConnection`
endpoints to take an `ErrorCodeCallbackHandler` instead of a
`CallbackHandler`. This callback type gets an `ErrorCode` instead of a
`bool` passed as the only parameter. This value should be compared to
`ErrorCode::None` to ensure the call succeeded.