[NT-0] fix: array constructor memory safety
The previous implementation of `csp::commn::array` included
support for constructing the array from a blob of memory.

The way it worked was via memcpy, which is unsafe if the memory
it is copying from describes non-PoD types that themselves hold
pointers to memory which they manage (such as strings).

This leads to behaviour where the dtor of one object invokes the
freeing of memory owned by another.

To address this, this change replaces the memcpy with a per-element
assignment, which makes use of the object's  assignment operator instead