[NT-0] doc: inline docs & correct error msg
From feedback, it has been found that there are valid use cases when
we need to express that an asset collection does not belong to a space.

So this change removes the error logging when this is detected.

Separately, this change also adds additional inline documentation to
AssetCollection.h, describing the purpose of each member of the
`AssetCollection` class (where it notes that if an asset collection does
 not belong to a space, the `SpaceID` member will be an empty string).