[OF-1257] fix: Exiting a space clears MP scopes
Now that the MP connection is live for the lifetime of the
application, and the scope for the space is set on invoking
`SpaceSystem::EnterSpace`, we must do the symmetric
opposite when exiting a space.

Since CSP enforces that a client can only listen to one scope
at a time at present, this is a case of invoking the `ResetScopes`
endpoint on the services when the client calls `ExitSpace`.