[OF-3308] refac!: refactored RefreshSession
* RefreshSession now internal function, returns a NullResult to
prevent confusion (should not be used for logging in) - functions
how it always did before, no MultiplayerConnection implementation.

* New function LoginWithRefreshToken - Renamed public API function
RefreshSession. It includes functionality to establish a multiplayer
connection as the assumption is this function is being used to log in
where there is still a valid refresh token to act on.