[NT-0] fix: UpdateSpace respects required fields
Certain fields on the services endpoint for updating spaces
are required, however we did not always provide them to the
endpoint.

Separately, this change also addresses a false expectation in a CSP
test, when a client does not provide a description string to update
space. The expected behaviour in this case is for the description to be
cleared by the services, and the test has been updated to respect this.