Staging 4.23.1 (#377)
* [NT-0] fix: Local entities are destroyed upon exiting a space

* [NT-0] fix: entity cleanup on space exit

As part of a recent broad refactor to the multiplayer connection's application lifetime, the connection now persists for the duration of the application's lifetime. Entering a space now narrows the connections focus to listen to entities within the space, and exiting stops the connection from listening to entities within the space.

A bug has arisen however where whilst leaving a space stops the application receiving updates regarding entities, it does not clear CSP's local view of currently known entities.

This change introduces a new function to SpaceEntitySystem, LocalDestroyAllEntities which is invoked by the SpaceSystem after it has finished listening for entities within the space. It clear's CSP's local view of all currently known entities.

Tests have been updated to validate this behaviour.

* [NT-0] fix: update services api (#375)

* [NT-0] fix: updating services API

A breaking change to Magnopus services requires an update
in order for CSP clients to update space descriptions

* [NT-0] fix: CSP code changes for latest services

Necessary code changes in order to adapt to latest
services API changes.

* [NT-0] fix: UpdateSpace respects required fields

Certain fields on the services endpoint for updating spaces
are required, however we did not always provide them to the
endpoint.

Separately, this change also addresses a false expectation in a CSP
test, when a client does not provide a description string to update
space. The expected behaviour in this case is for the description to be
cleared by the services, and the test has been updated to respect this.

* [NT-0] fix: update chs wrapper

---------

Co-authored-by: Sam Birley <sam.birley@magnopus.com>

---------

Co-authored-by: Sam Birley <99482500+MAG-SamBirley@users.noreply.github.com>
Co-authored-by: Sam Birley <sam.birley@magnopus.com>