[NT-0] fix: Error log message formatting on failure to disconnect when logging out (#387)
[NT-0] fix: addressing a log formatting bug

Discovered whilst debugging a client, when a client attempts to log out but does not have a valid MP connection, we emit an error message to the log. There was an error in the log formatting however, where it was expecting a string and being provided an int.

This change fixes the formatting issue, and also includes some updates to the 'OKOBuildTool' - a legacy script that remains highly useful to this day as a workflow enhancer when debugging CSP issues in Unreal client applications. The updates concern naming and paths that have changed since the tool was last updated.