[OF-1322] refac: Update generate shell scripts (#421)
As of Python 3.12+ it is a requirement that we use a venv to install
modules. The generate_solution_X.sh scripts for mac, ios and visionos
have all been updated to check for the version of Python installed,
and if >= 3.12, to create and activate a venv before proceeding with
module installation.