[OB-3568] fix: VisionOS unity build script
This change addresses an issue that was found to be preventing
the `libConnectedSpacesPlatform.a` binary from being included
in XCode projects generated by Unity.

The cause was an omission in a CSP-provided Unity build script,
where the logic that determines whether or not to include the binary
was not accounting for the VisionOS platform.