[NT-0] test: EventBrite data formatting update
EventBrite have made a services change that now
rejects data that does not match their expected schema.

We have addressed this in CSP tests by changing the data passed
into CSP to match the EventBrite schema.

We have also refactored the eticketing tests to all rely on a single
set of test variables, to simplify updates in the future.