[NT-0] feat: Space POIs returned by the POI system
The POI system now returns pois that have been associated
with a space. Previously there was an implicit search term
that filtered out any POIs that were denoted as space POIs.

The client application can determine whether a POI relates to a
space or not using the `Type` member variable held by the POI.