[NT-0] feat: Querying for local POIs by type (#450)
Discovered whilst writing spatial services documentation for CSP, there was a client-application-level feature missing, where whilst the client app is able to associate a POI with a space using the UpdateSpaceGeoLocation method on SpaceSystem, they had no affordance via the POISystem to later query for POIs associated with spaces.

In fact, the POISystem was deliberately filtering out POIs that were associated with spaces, and only ever returning DEFAULT type POIs.

This change introduces a new optional parameter to the POI System, Type, which allows the client to query for only POIs of a certain type.

If all types are desired to be included in the returned results, the optional value should not be specified.
If the previous behaviour of CSP when returning results is preferred, please pass in the DEFAULT type.
If the application wishes to filter for only spaces of type SPACE, please pass this as the argument.
The change also addresses some technical debt, where the strings denoting default and space type POIs were previously hard-coded in several methods. They have all been unified to rely on a single set of type/string conversion helper functions.