[NT-0] fix: crash on DLL load for CPP apps
This change concerns moving a static string declaration in the global
scope out of a public header. Specifically `SequenceHierarchyName`.

This causes CPP apps using CSP dlls to crash on load of the DLL as the
application attempts to instatiate the global-scope string variable on
load _before_ CSP has initialised its memory allocators (which the
string class depends upon).

Since the variable is only used within CSP, it has been moved to an
internal source file. To avoid creating a brand new header for a single
constant, the `MultiplayerKeyConstants.h` file has been used for this
and renamed to the more generic `MultiplayerConstants.h`.

The true fix though is to _not_ declare this variable at the global
scope, because this will always be risky with respect to CSP's
memory allocators.

Instead it is retrieved through a global-scope static function,
which more or less guarantees that it will only ever be called
after the allocators have been initialised.