[NT-0] fix: linking with new sequence constants
Addressed an issue where the new sequence constant
functions were both declared globally and static.

That leads to the functions technically only being
linkable within the current file, which is not what we want
(and was causing build failures).

Instead we wrap the functions inside a  dummy class and
continue declaring them static, which has the effect of allowing them
to be linked to from other files.