[NT-0] fix: crash on DLL load for CPP apps (#473)
Bug & Fix
This change concerns moving a static string declaration in the global scope out of a public header. Specifically SequenceHierarchyName.

Global variable strings within CSP can cause apps using CSP dlls to crash on load of the DLL, as it leads to the application attempting to instantiate the global-scope string variable on DLL-load, before CSP has initialised its memory allocators, which the string class depends upon. The end result is a null dereference crash.

The fix though is to not declare this string variable at the global scope, because this will always be risky with respect to CSP's memory allocators. Instead it is retrieved through a global-scope static function, which more or less guarantees that it will only ever be called after the allocators have been initialised.

Additionally...
Since in this case, since the variable is only used internally within CSP, it has also been moved to an internal source file.

To avoid creating a brand new header for a single constant, the MultiplayerKeyConstants.h file has been used for this and renamed to the more generic MultiplayerConstants.h.