[OF-1419] feat: sequence keys support spaces
CSP client applications can now reason about sequence
keys that include characters usually reserved in URIs,
such as spaces.

As part of this work, we have introduced a new set of encoding
source files for future encode/decode behavior.

During development, we discovered that there are some service
endpoints currently expecting the provided keys to be _doubly_
encoded. For now, we are rolling with this.