[OF-1407] feat Hotspot group CRUD events (#502)
The CSP API now includes surface area allowing the client to bind to a new event callback, via SetHotspotSequenceChangedCallback on the multiplayer connection.

The callback will fire when a hotspot group has a CRUD operation applied to it. It describes the change type (Create, Update, Rename, Delete), the hotspot group name and the ID of the space that the hotspot group relates to.