[NT-0] test: disabling hotspot group events
Since the order of multiplayer events is not guaranteed,
these checks can fail if the order does not come through
in the way the tests expect.

We plan to add additional logic to CSP to make the events
always appear to come through in a reliable order in the future,
but for we're disabling these checks to avoid flakey behavior.