[OB-3674] fix: Do not encode already encoded URLs
CSP made a change in March to encode URLs in asset details.
We added this encode earlier than that to workaround the issue for
a client team, but now it's causing double encode errors,
so we should remove it.

Added a couple of tests so we'd know if this changes again.