[NT-0] doc: wider content style
Addressing feedback from the OKO web team regarding width of main
content area. Increased content width from 800px to 13020px.

This has the effect of rendering the page symmetrically, where the
sidebar on the left is now equal in width to the negative space on
the right (on a 1920x1080 monitor).