[NT-0] doc: removing references to the GH wiki from the readme
With all GitHub wiki documentation having been added to the documentation site, this change replaces references to the wiki pages with CSP API docs links. 

It also restructures the readme to flow better from a first-time reader's point of view, and now includes the CSP logo instead of the emoji-based header.