[OF-1471] test: Integrate multiplayer test runner with Tests project (#575)
* [OF-1471] test: Add tiny-process-lib

Adds tiny-process-lib, for the purposes
of the multiplayer test runner, and
invoking it within the tests.

I laboured over making this a submodule vs
having it directly in thirdparty, and eventually
came to the conclusion that direct thirdparty
is more standard to the codebase.

* [OF-1471] test: Build & Install TPL

Build TPL using cmake during the solution creation.
Copy it to the directory of the tests so its easily
invokable.

We go to the effort to install tiny-process-lib to
an install directory here, which isn't something we
do for other cmake libs, but really we should.

* [OF-1471] test: Dont test MTR by default

During dev it seemed cute to have debug builds
run the tests by default, but it immediately became
frustrating during integration.

* [OF-1471] test: Test Runner Shakedowns

A couple of changes to the test runner that emerged
during integration.

1. I hit the classic problem of calling a private header
in a public one.

2. The process descriptors were too generic,
and I actually accidentlyoutput the same text during
debugging, which caused pain due to
how we monitor stdout to flag futures.

* [OF-1471] test: MTR process manager

Add the process controller object for use in our tests.

Allows setting command line arguments and invoking
the multiplayer test runner, and managing termination
via RAII.

The manager works via listening for set strings on
stdout, and then setting promises to notify the main
test of when events have occured on each process.

* [OF-1471] test: MTR integration test

Add an integration test that launches the
multiplayer test runner and checks that
its operations are performed.

* [OF-1471] test: Unit tests for runner process

Unit tests for the runner process object itself.
Not tests for the runner itself, which has its own tests,
(although they're not run automatically yet / ever)