[OF-1494] fix: Fix SpaceTransform multiplication
The existing implementation had several things wrong with it, one has to imagine
it was never called anywhere, because this would be so apparently.
Unfortunately, it was untested since the beginning.

Issues :
- [[nodiscard]] warning on normalize ignored, although post-rotation normalization is only needed to prevent cascading floating point errors
- The W component for glm::quat is the first argument, but for our vectors is is the last, all our quaternions were being constructed with nonsense values. (Strong typing please! Vector4 is inappropriate to store quaternions for precisely this reason)
- I don't think you can do complex compositions in this manner, separating the translation, scale and rotation components like that.
- Untested

The prior code, even with the code issues fixed, failed to pass even the identity assumptions in the new test suite that has been written.
For that reason, and due to the face that matrix compositions are more comprehensible, I've replaced the implementation with a matrix
based one. It's more verbose to be sure, and probably slightly slower than direct quaternion work, but at least it has a chance of being correct.