[NT-0] doc: Multiplayer architecture and events (#607)
This change introduces two new sections to the Spaces documentation module:

* Multiplayer Architecture
* Events

Note:

The change also introduces/demonstrates the pattern to be followed should it be helpful to provide the reader with a link to API documentation inline. For examples, please see references to eval_rst in Library\docs\source\learn\spaces\events.md.