[OF-1512] style: Add Async++
Add Async++, a library that provides in-thread continuations.

https://github.com/Amanieu/asyncplusplus

This is the attempt at solving the massive nested callbacks issue, this lib
will allow us to write callback based code linearly in the form of continuations.
This solution was came to after attempting more standard solutions, but all
of them require blocking in some form, which is forbidden in emscripten WASM.

Other options considered:
Boost.Threading - Gives the same capabilities, but has other boost dependencies
CppCoro - Maybe a slightly better solution, coroutine based, but requires C++20, and bills itself as "experimental"

Async++ is a minimal library that has existed for a while. Especially because of our limited WASM knowledge,
we should favour minimalism and stability wherever possible in our dependencies, so we have a chance to understand
them when something goes wrong.

Building it as a static library, so it works with WASM.

MIT Licensed. 1.4k Github stars.