[OF-1512] test: Parameterized EnterSpace Tests
Add a wide range of EnterSpace regression tests ahead of refactoring.
Add the capability to run googletest parameterized tests in our framework.

This has the minor downside in that each parameterized suite appears as it's own thing in the VS test explorer,
as it can't go more than 3 deep. Annoying, but more than worth it.