[OF-1512] refac: Refactor EnterSpace to be continuations based
Refactor the old nested callback version of EnterSpace into a continuation based one.

This defines a clear pipeline for entering a space.
- GetSpace
- ValidateSuccess
- AddUserToSpace(If Neccesary)
- ValidateSuccess
- EnterSpace
- RefreshScopes
- ValidateSuccess
- CleanupIfThrow

If you're aware of the concept, this sort of (monadic :o ) functional chaining is also known as railway oriented programming, there's a talk here that I like : https://vimeo.com/113707214
You may start to think that this could be even cleaner, without doing the explicit result checks in between all the steps and passing the Result type in an exception to the final handler. I think you're right. That would be a good improvement to make.

This doesn't refactor the refresh scopes functionality, which is moved into its own method. This method helpfully serves as an example as to how we can adapt existing nested callback code into chains without having to rewrite it all at once.

NOTE See the inline_scheduler() used everywhere. This is essential, it means that the continuations are run immediately, rather than being placed into a threadpool. This is what lets this work on WASM. It would be good to figure out how to set this as the default scheduler, you can do that.

NOTE Now that there's a clear place to reset state after a failed space entry, you may think that we're not resetting all the state. That's correct, we're not, (although we're resetting more now than we were before, there was an untracked bug with CurrentSpace that my regression testing found.) Once we sort out the scopes thing, we can maybe think about properly resetting state, but i'm not going crazy for now.