[OF-1512] refac: In-Thread Continuations, EnterSpace refactor. (#616)
Refactor the old nested callback version of EnterSpace into a continuation based one.

This defines a clear pipeline for entering a space.
- GetSpace
- ValidateSuccess
- AddUserToSpace(If Neccesary)
- ValidateSuccess
- EnterSpace
- RefreshScopes
- ValidateSuccess
- CleanupIfThrow

If you're aware of the concept, this sort of (monadic :o ) functional chaining is also known as railway oriented programming, there's a talk here that I like : https://vimeo.com/113707214
You may start to think that this could be even cleaner, without doing the explicit result checks in between all the steps and passing the Result type in an exception to the final handler. I think you're right. That would be a good improvement to make.

This doesn't refactor the refresh scopes functionality, which is moved into its own method. This method helpfully serves as an example as to how we can adapt existing nested callback code into chains without having to rewrite it all at once.

NOTE See the inline_scheduler() used everywhere. This is essential, it means that the continuations are run immediately, rather than being placed into a threadpool. This is what lets this work on WASM. It would be good to figure out how to set this as the default scheduler, you can do that.

NOTE Now that there's a clear place to reset state after a failed space entry, you may think that we're not resetting all the state. That's correct, we're not, (although we're resetting more now than we were before, there was an untracked bug with CurrentSpace that my regression testing found.) Once we sort out the scopes thing, we can maybe think about properly resetting state, but i'm not going crazy for now.

* [OF-1512] style: Add Async++

Add Async++, a library that provides in-thread continuations.

https://github.com/Amanieu/asyncplusplus

This is the attempt at solving the massive nested callbacks issue, this lib
will allow us to write callback based code linearly in the form of continuations.
This solution was came to after attempting more standard solutions, but all
of them require blocking in some form, which is forbidden in emscripten WASM.

Other options considered:
Boost.Threading - Gives the same capabilities, but has other boost dependencies
CppCoro - Maybe a slightly better solution, coroutine based, but requires C++20, and bills itself as "experimental"

Async++ is a minimal library that has existed for a while. Especially because of our limited WASM knowledge,
we should favour minimalism and stability wherever possible in our dependencies, so we have a chance to understand
them when something goes wrong.

Building it as a static library, so it works with WASM.

MIT Licensed. 1.4k Github stars.