[OB-3917] fix: change all int_64t exposed script parameters to int_32t
This is because our current version of quickjs does not support BigInt
(64 bits ints), which led to an error if using BigInt in a script on
Web, or an unsupported conversion warning if using a Number.

Now that we expose 32 bits ints, the web scripts no longer show an
unsupported conversion warning when using a Number.