[OF-1577] feat!: Add tag controls to Space, removing them from the metadata prototype. (#637)
* [OF-1577] chore: Update CSP Services

Update to v2.0-d20250224.rbdaf75117.Release
To expose the tags API.

* [OF-1566] feat: Add string `Contains` method

Needed for some tests later.

* [OF-1577] feat: Replace buffer construction with To/from Vector methods: Array

Needed later to sanely interop with DTO's, which sensibly use vectors.

* [OF-1577] feat: Add tags to SpaceDTO, Remove from Metadata callbacks

As title. Also migrated the local sets to use the new vector constructors, much clearer.

* [OF-1577] feat!: Alter public API for new tag modality

Change UpdateSpace and Metadata methods to now deal with tags on the space, rather than along with the metadata on the associated prototype.

Modify CreateSpace to set tags in the GroupInfo if they are set.
Remove all tag functionality from Metadata methods.

* [OF-1577] test: Add test for space filtering by tags functionality

Spaces can now be filtered by tags in various ways.
Also test UpdateSpace behaviour, as that's also new.

* [NT-0] chore: Add Chrono to spaceentitysystem

Must have been compiling all this time as a consequence of transitive includes.

It failed to compile for me, and chrono is for sure used, check below in the header, system_clock is used.

* fixup! [OF-1577] test: Add test for space filtering by tags functionality

Use unique space tags for test

* [OF-1577] style: Revert to using pre-existing convert functions

Not sure this is better, but it's a toss up and some members of the
team prefer it.

* Empty Commit for CI

---------

Co-authored-by: Sam Birley <99482500+MAG-SamBirley@users.noreply.github.com>