[OF-1575] feat: Pass endpoint to mulitplayer test runner
Pass the active endpoint to the multiplayer test runner, so it no longer uses its fallback behaviour.
Also add some logging to the runner, initially we were trying not to pollute stdout, but in the grand scheme of things, it's a small price to pay to get a bit more feedback.