[OF-1575] feat: Allow multiplayer service to work with specified URIs
The multiplayer connection was not at all ready to take arbitrary URIS.

This commit does the work to allow this to happen, it's mostly concerned with rationalizing (somewhat) the URL transformation we do.
It's still not ideal, URL/URI processing is a complicated topic.

As some of this logic is done in some gnarly stateful methods, I have made some effort to try and test the new logic, despite not being able to test the method as a whole :(

Not how this transformation is done in POCO, this means local-chs WILL NOT WORK ON WASM YET. This isn't a requirement for right this second, but the behaviour variance will be a problem, and we'll need it when we add wasm testing, so we shouldn't allow this to become the norm for too long.