[OF-1575] Enable Local CHS, Remove hard-coded URIs, migrate multiplayer URI handling. (#649)
* [OF-1575] feat: Pass endpoint to mulitplayer test runner

Pass the active endpoint to the multiplayer test runner, so it no longer uses its fallback behaviour.
Also add some logging to the runner, initially we were trying not to pollute stdout, but in the grand scheme of things, it's a small price to pay to get a bit more feedback.

* [OF-1575] feat: Remove remaining hard coded URLS

* [OF-1575] feat: Remove unnecesary URL restrictions in tests

We were checking conformance for URIS that isn't neccesary, as we want to allow localhost style URIS

* [OF-1575] test: Get the tests working that I could figure out

These were the easy tests that were failing, due to the known CHS processing delay on re-entering a space, as well as one issue where the callback was set too late.

* [OF-1575] feat: Allow multiplayer service to work with specified URIs

The multiplayer connection was not at all ready to take arbitrary URIS.

This commit does the work to allow this to happen, it's mostly concerned with rationalizing (somewhat) the URL transformation we do.
It's still not ideal, URL/URI processing is a complicated topic.

As some of this logic is done in some gnarly stateful methods, I have made some effort to try and test the new logic, despite not being able to test the method as a whole :(

Not how this transformation is done in POCO, this means local-chs WILL NOT WORK ON WASM YET. This isn't a requirement for right this second, but the behaviour variance will be a problem, and we'll need it when we add wasm testing, so we shouldn't allow this to become the norm for too long.

* fixup! [OF-1575] feat: Allow multiplayer service to work with specified URIs

Follow doc standards

* fixup! [OF-1575] feat: Allow multiplayer service to work with specified URIs

Catch exception in Start method