[OF-1585] refac: Inject SignalR connection
Refactor to allow SignalRConnection to be injected to `Connect`.
This has (or will have, in upcoming commits) the benefit of allowing us to fully mock SignalRConnection, in order to test network issues.

This is done by defining an interface that the concrete SignalRConnection can extend from, but also a mock type can.