[OF-1585] refac: Parse signalR errors to retain error message
This is pretty unfortunate, we go to a lot of effort to do effectively nothing and convert signalR exceptions into error codes. In a fully refactored multiplayer system, i'd like to see this rethought.

In this work, we need to provide reasons why errors happen, especially because ParseError can only give two return values (SpaceUserLimitExceeded or Unknown). Often we just don't provide any context, this attempts to at least forward through the reason the original exception is thrown.