[OF-1585] test: Do not send SignalR invocation during test teardown
There's no need to do this, we do it in each test case setup anyway. The original commit seemed to just put it here for safety?

It is problematic when you've set a mocked SignalRClient, as it gets destructed before the test teardown happens.

Alternately, we could instantiate the mocks in the fixture setup (a good reason to avoid our overly constrictive macros), but since this is unnecesary, just delete these lines.