[OF-1585] refac: Connect Continuations (#651)
* [OF-1595] refac: Pass exception to exception handler continuation

This change needs to parse the exception .what() in order to generate an error code.
This is a pretty unfortunate and verbose way of doing error handling imo, but the continuation change is good by itself.

* [OF-1585] refac: Rename continuation utility functions -> HTTP to be clearer

As there are two categories of error, HTTP and Multiplayer, be clear that the utility function written so far are for the HTTP ones.

* [OF-1585] refac: Inject SignalR connection

Refactor to allow SignalRConnection to be injected to `Connect`.
This has (or will have, in upcoming commits) the benefit of allowing us to fully mock SignalRConnection, in order to test network issues.

This is done by defining an interface that the concrete SignalRConnection can extend from, but also a mock type can.

* [OF-1585] refac: Add HTTPHeaders function

To allow mocking, add HTTP headers function that can be mocked, rather than direct private member access (extremely naughty!)

* [OF-1585] refac: Parse signalR errors to retain error message

This is pretty unfortunate, we go to a lot of effort to do effectively nothing and convert signalR exceptions into error codes. In a fully refactored multiplayer system, i'd like to see this rethought.

In this work, we need to provide reasons why errors happen, especially because ParseError can only give two return values (SpaceUserLimitExceeded or Unknown). Often we just don't provide any context, this attempts to at least forward through the reason the original exception is thrown.

* [OF-1585] refac: Use mock to test failing network state in Connect

Use a mock, combined with the injectable interface, to test the conditions where the signalR connection errors, simulating a permissions error, or a network drop, or whatever.

* [OF-1585] test: Do not send SignalR invocation during test teardown

There's no need to do this, we do it in each test case setup anyway. The original commit seemed to just put it here for safety?

It is problematic when you've set a mocked SignalRClient, as it gets destructed before the test teardown happens.

Alternately, we could instantiate the mocks in the fixture setup (a good reason to avoid our overly constrictive macros), but since this is unnecesary, just delete these lines.

* [OF-1585] refac: Refactor Connect to Continuations

Refactor Connect to Continuations.

As all the SignalR interaction is callback based itself, we use events heavily to paper the seams between the SignalR callbacks and our callbacks.

Errors are handled via exception, in the final continuation, meaning all errors will have uniform reporting.

* fixup! [OF-1585] refac: Use mock to test failing network state in Connect

Use CSP_NEW to avoid (sporadic ... weirdly) memory cleanup errors. This is annoying

* [OF-1585] test: Change expected connection message in Eventbus tests

* fixup! [OF-1585] refac: Refactor Connect to Continuations

Fix trailing S

* fixup! [OF-1585] refac: Refactor Connect to Continuations

Remove in progress comment