[OF-1586] refac: Remove unnecesary dynamic memory from SpaceEntity
This was done because I thought SpaceEntity needed to be movable, but it doesn't. Nonetheless, this is a better thing to be doing.

This makes mutexes In SpaceEntity mutable, which they already implicitly were anyway because we were using the pointer indirection to break const.  Better this way imo where it's explicit.