[OF-1586] feat: CreateAvatar continuations
Transform CreateAvatar to be continuations based.
The flow is

- Get the avatar ID from the server (RemoteGenerateNewAvatarId)
- Send the new avatar to the server (SendNewAvatarObjectMessage)
- Create the avatar locally (CreateNewLocalAvatar)

You'll see I've gone to some effort to prevent the possibility of leaks on error. This involves instantiating the avatar twice, (although not doing the network interaction twice, as we reuse the prior continuation result via the when_all branch) as we can't properly thread shared memory through a split continuation unless we've somewhere to put it at the end, which would have been too large a refactor. See comment in code.

Not sure about the specific continuation names, or even if I've quite got the concept communication right (especially for SendNewAvatarObjectMessage) open to change.